Move the merge_variables and move_variables functions to the process module since they aren't one-in, one-out
