+++ clarite/modules/process.py
@@ -12,11 +12,19 @@ Functions used to process data from one form into another, such as categorizing 
      :toctree: modules/process
 
      categorize
+     merge_variables
+     move_variables
 
 """
 
+from typing import List, Optional
 
-def categorize(data, cat_min: int = 3, cat_max: int = 6, cont_min: int = 15):
+import pandas as pd
+
+from ..internal.utilities import _validate_skip_only
+
+
+def categorize(data: pd.DataFrame, cat_min: int = 3, cat_max: int = 6, cont_min: int = 15):
     """
     Divide variables into binary, categorical, continuous, and ambiguous dataframes
 
@@ -97,3 +105,72 @@ def categorize(data, cat_min: int = 3, cat_max: int = 6, cont_min: int = 15):
     other_df = data.loc[:, other_filter]
 
     return bin_df, cat_df, cont_df, other_df
+
+
+def merge_variables(data: pd.DataFrame, other: pd.DataFrame, how: str = 'outer'):
+    """
+    Merge a list of dataframes with different variables side-by-side.  Keep all observations ('outer' merge) by default.
+
+    Parameters
+    ----------
+    data: pd.Dataframe
+        "left" DataFrame
+    other: pd.DataFrame
+        "right" DataFrame which uses the same index
+    how: merge method, one of {'left', 'right', 'inner', 'outer'}
+        Keep only rows present in the left data, the right data, both datasets, or either dataset.
+
+    Examples
+    --------
+    >>> import clarite
+    >>> df = clarite.modify.merge_variables(df_bin, df_cat, how='outer')
+    """
+    return data.merge(other, left_index=True, right_index=True, how=how)
+
+
+def move_variables(data: pd.DataFrame, other: pd.DataFrame,
+                   skip: Optional[List[str]] = None, only: Optional[List[str]] = None):
+    """
+    Move one or more variables from one DataFrame to another
+
+    Parameters
+    ----------
+    data: pd.Dataframe
+        DataFrame containing the variable(s) to be moved
+    other: pd.DataFrame
+        DataFrame (which uses the same index) that the variable(s) will be moved to
+    variables: List of strings
+        Names of the variables to move (all are moved by def)
+
+    Returns
+    -------
+    data: pd.DataFrame
+        The first DataFrame with the variables removed
+    other: pd.DataFrame
+        The second DataFrame with the variables added
+
+    Examples
+    --------
+    >>> import clarite
+    >>> df_cat, df_cont = clarity.process.move_variables(df_cat, df_cont, only=["DRD350AQ", "DRD350DQ", "DRD350GQ"])
+    Moved 3 variables.
+    >>> discovery_check, discovery_cont = clarite.process.move_variables(discovery_check, discovery_cont)
+    Moved 39 variables.
+    """
+    # Which columns
+    columns = _validate_skip_only(list(data), skip, only)
+
+    # Add to new df
+    other = merge_variables(other, data[columns])
+
+    # Remove from original
+    data = data.drop(columns, axis='columns')
+
+    # Log
+    if len(columns) == 1:
+        print("Moved 1 variable.")
+    else:
+        print(f"Moved {len(columns)} variables.")
+    
+    # Return
+    return data, other
